renamed test file to be more general, and added another test
